requirements.txt
asgiref==3.11.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cssselect2==0.8.0
Django==5.2.9
django-cloudinary-storage==0.3.0
django-nested-admin==4.1.6
fonttools==4.61.1
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.1
python-monkey-business==1.1.0
requests==2.32.5
resend==2.19.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
urllib3==2.6.2
weasyprint==67.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
